rfc: change file structure
