+++ package.json
@@ -7,6 +7,7 @@
         "@testing-library/jest-dom": "^5.17.0",
         "@testing-library/react": "^13.4.0",
         "@testing-library/user-event": "^13.5.0",
+        "axios": "^1.6.2",
         "craco-alias": "^3.0.1",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
@@ -16,10 +17,10 @@
         "web-vitals": "^2.1.4"
     },
     "scripts": {
-        "start": "react-scripts start",
-        "build": "react-scripts build",
-        "test": "react-scripts test",
-        "eject": "react-scripts eject",
+        "start": "craco start",
+        "build": "craco build",
+        "test": "craco test",
+        "eject": "craco eject",
         "commitlint": "npx commitlint --edit",
         "lint-check": "npx eslint .",
         "lint-fix": "npx eslint . --fix",
