+++ src/App.jsx
@@ -1,10 +1,10 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Head from './components/Head';
-import Landing from './pages/Landing';
-import FindPw from './pages/FindPw';
-import RegisterCommon from './pages/RegisterCommon';
-import RegisterInstitution from './pages/RegisterInstitution';
-import RegisterIndividual from './pages/RegisterIndividual';
+import { Head } from '@components';
+import Landing from '@pages/Landing';
+import FindPw from '@pages/FindPW';
+import RegisterCommon from '@pages/Register/RegisterCommon';
+import RegisterInstitution from '@pages/Register/RegisterInstitution';
+import RegisterIndividual from '@pages/Register/RegisterIndividual';
 
 function App() {
     return (
