+++ src/apis/landing/index.js
@@ -1,0 +1,1 @@
+import axios from 'axios';
