+++ src/components/About/index.jsx
@@ -1,0 +1,10 @@
+import { Background, Container, Title } from './style';
+export const About = () => {
+    return (
+        <Background>
+            <Container>
+                <Title></Title>
+            </Container>
+        </Background>
+    );
+};
