+++ src/components/About/style.js
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+
+export const Background = styled.div`
+    height: 100vh;
+    overflow: hidden;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+`;
+export const Container = styled.div`
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+`;
+export const Title = styled.h1`
+    font-size: 40px;
+`;
