+++ src/components/Head/style.js
@@ -1,8 +1,7 @@
 import styled from 'styled-components';
 import { NavLink } from 'react-router-dom';
-import { Outlet } from 'react-router-dom';
 
-const Header = styled.div`
+export const Header = styled.div`
     display: flex;
     padding: 1.2rem;
     width: 1180px;
@@ -15,7 +14,7 @@ const Header = styled.div`
     }
 `;
 
-const IVent = styled(NavLink)`
+export const IVent = styled(NavLink)`
     font-size: 2rem;
     font-family: 'Pretendard-Bold', sans-serif;
     cursor: pointer;
@@ -27,18 +26,3 @@ const IVent = styled(NavLink)`
         color: rgb(134, 38, 51);
     }
 `;
-
-const Head = () => {
-    return (
-        <>
-            <Header>
-                <IVent to='/'>iVent</IVent>
-            </Header>
-            <main>
-                <Outlet />
-            </main>
-        </>
-    );
-};
-
-export default Head;
