+++ src/components/Successfulvents/index.jsx
@@ -1,0 +1,8 @@
+import { Container, Title } from './style';
+export const SuccessfulIvents = () => {
+    return (
+        <Container>
+            <Title>#iVents #성공적</Title>
+        </Container>
+    );
+};
