+++ src/components/Upcominglvents/index.jsx
@@ -1,0 +1,8 @@
+import { Container, Title } from './style';
+export const UpcomingIvents = () => {
+    return (
+        <Container>
+            <Title>#인기 #iVents</Title>
+        </Container>
+    );
+};
