+++ src/components/Upcominglvents/style.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+`;
+
+export const Title = styled.h1`
+    font-size: 40px;
+`;
