+++ src/components/index.js
@@ -1,0 +1,5 @@
+export * from './About';
+export * from './Head';
+export * from './Login';
+export * from './Successfulvents';
+export * from './Upcominglvents';
