+++ src/pages/Landing/style.js
@@ -1,0 +1,49 @@
+import styled from 'styled-components';
+
+export const TopContainer = styled.div`
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+`;
+
+export const Container1 = styled.div`
+    display: flex;
+    height: 400px;
+`;
+
+export const Left1 = styled.div`
+    flex: 3;
+    background-color: #eee;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+`;
+
+export const Right1 = styled.div`
+    flex: 2;
+    background-color: #ddd;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+`;
+
+export const Container2 = styled.div`
+    display: flex;
+    justify-content: space-between;
+    width: 1180px;
+    margin: 0 auto;
+`;
+
+export const Left2 = styled.div`
+    flex: 1;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+`;
+
+export const Right2 = styled.div`
+    flex: 1;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+`;
