+++ src/pages/Register/RegisterCommon/style.js
@@ -1,29 +1,29 @@
 import styled from 'styled-components';
 import { NavLink } from 'react-router-dom';
 
-const Background = styled.div`
+export const Background = styled.div`
     height: 100vh;
     overflow: hidden;
     display: flex;
     justify-content: center;
     align-items: center;
 `;
-const Container = styled.div`
+export const Container = styled.div`
     display: flex;
     flex-direction: column;
     justify-content: center;
     margin-top: -100px;
 `;
 
-const Title = styled.h1`
+export const Title = styled.h1`
     font-size: 50px;
     text-align: center;
 `;
-const ButtonContainer = styled.div`
+export const ButtonContainer = styled.div`
     display: flex;
     justify-content: space-between;
 `;
-const Button = styled(NavLink)`
+export const Button = styled(NavLink)`
     margin: 120px;
     margin-top: 40px;
     padding: 20px;
@@ -39,19 +39,3 @@ const Button = styled(NavLink)`
         border: 1px solid #ccc;
     }
 `;
-
-const RegisterCommon = () => {
-    return (
-        <Background>
-            <Container>
-                <Title>iVent 사용 목적을 알려주세요!</Title>
-                <ButtonContainer>
-                    <Button to='/registerinstitution'>기관/단체</Button>
-                    <Button to='/registerindividual'>개인</Button>
-                </ButtonContainer>
-            </Container>
-        </Background>
-    );
-};
-
-export default RegisterCommon;
