+++ src/pages/Register/index.js
@@ -1,0 +1,3 @@
+export * from './RegisterCommon';
+export * from './RegisterIndividual';
+export * from './RegisterInstitution';
