+++ src/pages/index.js
@@ -1,0 +1,4 @@
+export * from './FindPW';
+export * from './Landing';
+// export * from './MyPage';
+export * from './Register';
